==> nb/1.build_data_sources/1.find_modified_concepts.ipynb <==
library(tidyverse)

modifiers <- 'severe|significant|serious'

# These files have rows with entries that start with / " and other problematic
#  characters. Luckily there are no ~ used in either, so I chose this as the
#  quote char to avoid read errors.
concept <- read_tsv('../data/athena/CONCEPT.csv', quote="~") %>% 
    filter(vocabulary_id == 'SNOMED')

synonym <- read_tsv('../data/athena/CONCEPT_SYNONYM.csv', quote="~") %>% 
    filter(language_concept_id == 4180186) %>%  # Only keep English concepts
    select(-language_concept_id)

relationship <- read_tsv('../data/athena/CONCEPT_RELATIONSHIP.csv', quote="~")

ancestor <- read_tsv('../data/athena/CONCEPT_ANCESTOR.csv', quote="~")

# Concept names vs synonyms

We want concepts that have either one of the modifiers in the `concept_name` or in the name of a synonym. Morbid obesity is a good example of this.

| concept_id |	concept_name |	domain_id |	concept_code |	concept_synonym_name |
| -- |	-- |	-- |	-- |	-- |
| 40565487 |	Morbid obesity |	Condition	| 389986000 | OBESITY, SEVERE |
| 40565487 |	Morbid obesity |	Condition	| 389986000 | Severe obesity |
| 40565487 |	Morbid obesity |	Condition	| 389986000 | obesity severe |
| 40565487 |	Morbid obesity |	Condition	| 389986000 | severe obesity	 |

A couple more good examples are below.

| concept_id |	concept_name |	domain_id |	concept_code |	concept_synonym_name |
| -- |	-- |	-- |	-- |	-- |
| 440370 |	Nutritional marasmus |	Condition |	29740003 |	Severe malnutrition |
| 256716 |	Asthma with status asthmaticus |	Condition | 	57546000 |	acute severe asthma	 |

concept_with_modifiers <- concept %>% 
    filter(concept_class_id == 'Clinical Finding') %>%
    filter(domain_id %>% str_detect('Condition')) %>%    
    select(concept_id, concept_name, concept_code) %>%

    # Want all concepts that have a synonym with a modifier
    left_join(synonym, by = 'concept_id') %>%

    # For some reason, not all concept_names are given as synonyms themselves
    #  in the CONCEPT_SYNONYM table. Have to check both separately.
    mutate(
        name_has_mod = concept_name %>% str_to_lower %>% str_detect(modifiers),
        syn_has_mod = concept_synonym_name %>% str_to_lower %>% str_detect(modifiers),
    ) %>%
    
    # All concepts with either a synonym or name having a modifier
    filter(name_has_mod | syn_has_mod) %>%

    # Just unique concepts, so we can find parents (HLT)
    select(concept_id, concept_name, concept_code) %>%
    distinct() %>%
    drop_na()

concept_with_modifiers %>% write_tsv('../data/computed/concepts_with_modifiers.tsv')

concept_with_modifiers %>% head(2)

## Parents of modified concepts

Rather than a top-down approach, in which I try to pick a distance from the root of SNOMED diseases, say, I think it is better to just look at the parents of modified concepts.

First find the parents of modified children, then find synonyms of those parents. Finally, associate each parent with ALL its descendants.

parents_modified_children <- concept_with_modifiers %>%
    left_join(
        relationship %>% 
            filter(relationship_id == 'Is a'), 
        by = c('concept_id' = 'concept_id_1')
     ) %>%
    select(
        modified_concept_id = concept_id,
        modified_concept_name = concept_name,
        modified_concept_code = concept_code,
        parent_concept_id = concept_id_2,
    ) %>%
    left_join(
        concept %>% 
        select(concept_id, parent_concept_name = concept_name, 
               parent_concept_code = concept_code), 
        by = c('parent_concept_id' = 'concept_id')
    )  %>%
    drop_na()

parents_modified_children %>% write_tsv('../data/computed/parents_with_modified_children.tsv')

parents_modified_children %>% head(2)

parents_synonyms <- parents_modified_children %>%
    select(starts_with('parent')) %>%
    left_join(synonym, by = c('parent_concept_id' = 'concept_id')) %>%
    drop_na()

parents_synonyms %>% write_tsv('../data/computed/parents_synonyms.tsv')

parents_synonyms %>% head(2)

parent_to_descendant_synonyms <- parents_modified_children %>%
    distinct(parent_concept_id) %>%
    left_join(ancestor, by = c('parent_concept_id' = 'ancestor_concept_id')) %>%
    select(parent_concept_id, descendant_concept_id) %>%
    left_join(synonym, by = c('descendant_concept_id' = 'concept_id')) %>%
    rename(descendant_synonym_name = concept_synonym_name) %>%
    drop_na()

parent_to_descendant_synonyms %>% write_tsv('../data/computed/parent_to_descendant_synonyms.tsv')

parent_to_descendant_synonyms %>% head(2)

parent_to_desc_syn_codes <- parent_to_descendant_synonyms %>%
    inner_join(
        concept %>% 
            select(concept_id, concept_code),
        by = c('parent_concept_id' = 'concept_id')
    ) %>%
    select(parent_concept_code = concept_code, descendant_concept_id, descendant_synonym_name) %>%
    inner_join(
        concept %>% 
            select(concept_id, concept_code),
        by = c('descendant_concept_id' = 'concept_id')
    ) %>% 
    select(parent_concept_code, descendant_concept_code = concept_code, descendant_synonym_name)

parent_to_desc_syn_codes %>% write_tsv('../data/computed/parent_to_descendant_synonyms_codes.tsv')

parent_to_desc_syn_codes %>% head(2)
